update tên file cmm014

+++ nts.uk/web/nts.uk.pr.web/src/main/webapp/view/cmm/014/a/cmm014.a.vm.js
@@ -126,13 +126,16 @@ var cmm014;
                     if (self.dataSource().length > 0) {
                         var item = new model.RemoveClassificationCommand(self.currentItem().classificationCode);
                         self.index_of_itemDelete = self.dataSource().indexOf(self.currentItem());
-                        a.service.removeClassification(item).done(function (res) {
-                            self.getClassificationList_aftefDelete();
-                        }).fail(function (res) {
-                            if (res.message == "ER06") {
-                                alert("対象データがありません。");
-                            }
-                            dfd.reject(res);
+                        nts.uk.ui.dialog.confirm("データを削除します。\r\nよろしいですか？").ifYes(function () {
+                            a.service.removeClassification(item).done(function (res) {
+                                self.getClassificationList_aftefDelete();
+                            }).fail(function (res) {
+                                if (res.message == "ER06") {
+                                    alert("対象データがありません。");
+                                }
+                                dfd.reject(res);
+                            });
+                        }).ifNo(function () {
                         });
                     }
                     else {
